stmhal: Renamed startup/system/_it.[ch] file to generic names
