+++ stmhal/startup_stm32.S
