+++ stmhal/stm32_it.c
@@ -67,7 +67,7 @@
 
 #include <stdio.h>
 
-#include "stm32f4xx_it.h"
+#include "stm32_it.h"
 #include STM32_HAL_H
 
 #include "py/obj.h"
